requirements.txt
# refer to: https://stackoverflow.com/questions/61741997/how-to-format-requirements-txt-when-package-source-is-from-specific-websites
# [warning] When `deepspeed` release new version, different workers might install different versions, and it will cause
# problems like this:: some worker got stuck at the deepspeed initialization stage, while others are not
tensorboardX
hydra-core>=1.3.2
omegaconf>=2.3.0
deepspeed==0.15.4
triton
torcheval
torchmetrics
torch_geometric
ogb==1.3.6
rdkit==2023.3.3
networkx[default]==3.1
urllib3
numpy<2
sentencepiece

# FlashAttention is only supported on CUDA 11.6 and above
transformers==4.53.3
timm
einops
# flash-attn<=2.4.2

huggingface-hub==0.30.2
accelerate==1.0.1
